Fixed printing `size_t` values on MinGW.

+++ SpinningIndicator.c
@@ -11,7 +11,7 @@
 uint64_t GetTickCounts() {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
-	return (tv.tv_sec * 1000 + tv.tv_usec / 1000);
+	return (uint64_t) (tv.tv_sec * 1000 + tv.tv_usec / 1000);
 }
 #endif // __GNUC__
